+++ sdk/core/CHANGELOG.md
@@ -1,5 +1,6 @@
 # release 5.0.1 (2021-06-18)
   * Improved type inference
+  * `CommonNetworkMember._affinity` made protected
 # release 5.0.0 (2021-06-16)
   * Implemented `WalletStorageService.fetchAllBlobs`.
   * Fixed `WalletStorageServices.fetchAllEncryptedCredentialsInBatches` stopping on batches with deleted entries.